+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: dayano <dayano@student.42.fr>              +#+  +:+       +#+         #
+#    By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/04/03 12:55:20 by ttsubo            #+#    #+#              #
-#    Updated: 2025/04/29 16:48:58 by dayano           ###   ########.fr        #
+#    Updated: 2025/05/01 14:29:53 by ttsubo           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -57,7 +57,7 @@ PARSER_SRC		=	allocate_cmds.c  parser.c  parser_utils.c  setup_cmds.c \
 					expand_env.c
 BUILTIN_SRC		=	cd.c exit.c pwd.c echo.c env.c unset.c export.c \
 					env_utils.c env_utils_2.c env_utils_3.c builtin_utils.c \
-					export_exec.c export_print_sorted_env.c export_error.c
+					export_exec.c export_print_sorted_env.c export_error.c split_key_value.c
 
 # 3.Add more directories as they are added.
 $(eval $(call add_module,root))
